+++ logger.html
@@ -631,3 +631,83 @@ info: GET es /static/js/crearTicket.js - 00:12:45
 info: GET es /static/home - 00:12:49
 info: GET es /static/css/home.css - 00:12:50
 info: GET es /static/js/home.js - 00:12:50
+info: GET es /static/login - 00:17:06
+info: GET es /static/js/logIn.js - 00:17:06
+info: GET es /static/css/session.css - 00:17:06
+info: POST es /api/sessions/login - 00:17:28
+info: GET es /static/home - 00:17:32
+info: GET es /js/home.js - 00:17:33
+info: GET es /css/home.css - 00:17:33
+info: GET es /static/js/home.js - 00:17:34
+info: GET es /static/css/home.css - 00:17:34
+info: GET es /static/home - 00:17:41
+info: GET es /css/home.css - 00:17:43
+info: GET es /static/css/home.css - 00:17:43
+info: GET es /static/js/home.js - 00:17:43
+info: GET es /js/home.js - 00:17:43
+info: POST es /api/cart/65c198277876613ca53269f9/purchase - 00:17:47
+info: GET es /static/ticket - 00:17:49
+info: GET es /static/js/crearTicket.js - 00:17:50
+info: GET es /static/css/ticket.css - 00:17:50
+info: GET es /static/home - 00:17:54
+info: GET es /js/home.js - 00:17:55
+info: GET es /static/css/home.css - 00:17:55
+info: GET es /css/home.css - 00:17:55
+info: GET es /static/js/home.js - 00:17:55
+info: OPTIONS es /assets/icon.svg - 00:23:51
+info: GET es /assets/icon.svg - 00:23:51
+info: GET es /static/login - 11:17:02
+info: GET es /static/css/session.css - 11:17:03
+info: GET es /static/js/logIn.js - 11:17:04
+info: GET es /static/login - 11:17:33
+info: GET es /static/css/session.css - 11:17:33
+info: GET es /static/js/logIn.js - 11:17:35
+info: POST es /api/sessions/login - 11:17:45
+info: GET es /static/css/session.css - 11:17:52
+info: GET es /static/home - 11:17:53
+info: GET es /css/home.css - 11:17:55
+info: GET es /static/js/home.js - 11:17:56
+info: GET es /static/css/home.css - 11:17:57
+info: GET es /js/home.js - 11:17:57
+info: POST es /api/cart/65c198277876613ca53269f9/purchase - 11:18:19
+info: GET es /static/ticket - 11:18:24
+info: GET es /static/js/crearTicket.js - 11:18:25
+info: GET es /static/css/ticket.css - 11:18:25
+info: GET es /static/home - 11:18:29
+info: GET es /js/home.js - 11:18:31
+info: GET es /css/home.css - 11:18:31
+info: GET es /static/css/home.css - 11:18:31
+info: GET es /static/js/home.js - 11:18:32
+info: GET es /static/logOut - 11:18:37
+info: GET es /static/css/session.css - 11:18:38
+info: GET es /static/js/logOut.js - 11:18:38
+info: GET es /api/sessions/logout - 11:18:40
+info: GET es /static/login - 11:18:40
+info: GET es /static/css/session.css - 11:18:41
+info: GET es /static/js/logIn.js - 11:18:41
+info: POST es /api/sessions/login - 11:18:53
+info: GET es /static/home - 11:19:02
+info: GET es /js/home.js - 11:19:03
+info: GET es /css/home.css - 11:19:03
+info: GET es /static/css/home.css - 11:19:03
+info: GET es /static/js/home.js - 11:19:03
+info: GET es /static/home - 11:19:42
+info: GET es /css/home.css - 11:19:44
+info: GET es /js/home.js - 11:19:44
+info: GET es /static/js/home.js - 11:19:44
+info: GET es /static/css/home.css - 11:19:44
+info: GET es /static/logOut - 11:19:46
+info: GET es /static/css/session.css - 11:19:47
+info: GET es /static/js/logOut.js - 11:19:47
+info: GET es /api/sessions/logout - 11:19:49
+info: GET es /static/login - 11:19:50
+info: GET es /static/css/session.css - 11:19:50
+info: GET es /static/js/logIn.js - 11:19:50
+info: POST es /api/sessions/login - 11:20:21
+info: GET es /static/home - 11:20:31
+info: GET es /js/home.js - 11:20:34
+info: GET es /static/css/home.css - 11:20:34
+info: GET es /css/home.css - 11:20:34
+info: GET es /static/js/home.js - 11:20:34
+info: POST es /api/cart/65c198277876613ca53269f9/purchase - 11:33:06
+info: POST es /api/cart/65c198277876613ca53269f9/purchase - 11:34:56
